Remove conflicting swift binding
In an attempt to package gdscript-formatter for NixOS[^1], we've
encountered an issue with a dependency hash conflict for Darwin.

According to some investigations[^2]:

> gdscript-formatter depends on a fork of tree-sitter-gdscript, which
has both the `bindings/swift/TreeSitterGdscript` directory and
`bindings/swift/TreeSitterGDscript` directory
>
> since macos is case insensitive, this is not possible and the files
involved forced to overwrite each other

Thus producing a different hash between Darwin and Linux.

---

This PR removes the extra directory, which fixes the issue.

[^1]: https://github.com/NixOS/nixpkgs/pull/445935
[^2]: https://github.com/NixOS/nixpkgs/pull/445935#discussion_r2381724211